server/requirements.txt
click==8.1.7
Flask==2.2.2
Flask-Cors==4.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
KrovetzStemmer==0.8
MarkupSafe==2.1.5
nltk==3.8.1
packaging==23.2
python-dotenv==1.0.1
rake-nltk==1.0.6
regex==2023.12.25
tqdm==4.66.2
Werkzeug==2.2.2
